chore: move analytics redux branch to ide
