+++ app/client/src/actions/ideActions.ts
@@ -35,3 +35,22 @@ export const setShowQueryCreateNewModal = (payload: boolean) => {
     payload,
   };
 };
+
+export const recordAnalyticsForSideBySideWidgetHover = (
+  widgetType: string,
+) => ({
+  type: ReduxActionTypes.RECORD_ANALYTICS_FOR_SIDE_BY_SIDE_WIDGET_HOVER,
+  payload: widgetType,
+});
+
+export const sendAnalyticsForSideBySideHover = () => ({
+  type: ReduxActionTypes.SEND_ANALYTICS_FOR_SIDE_BY_SIDE_HOVER,
+});
+
+export const recordAnalyticsForSideBySideNavigation = () => ({
+  type: ReduxActionTypes.RECORD_ANALYTICS_FOR_SIDE_BY_SIDE_NAVIGATION,
+});
+
+export const resetAnalyticsForSideBySideHover = () => ({
+  type: ReduxActionTypes.RESET_ANALYTICS_FOR_SIDE_BY_SIDE_HOVER,
+});
