+++ app/client/src/layoutSystems/common/AnalyticsWrapper/AnalyticsWrapper.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { LAYOUT_WRAPPER_ID } from "./constants";
-import { sendAnalyticsForSideBySideHover } from "actions/analyticsActions";
+import { sendAnalyticsForSideBySideHover } from "actions/ideActions";
 import { getLayoutSystemType } from "selectors/layoutSystemSelectors";
 import { LayoutSystemTypes } from "layoutSystems/types";
 import { useIsInSideBySideEditor } from "IDE/hooks";
@@ -10,23 +10,27 @@ export const AnalyticsWrapper: React.FC = ({ children }) => {
   const dispatch = useDispatch();
   const isInSideBySideEditor = useIsInSideBySideEditor();
   const layoutSystemType = useSelector(getLayoutSystemType);
+  const isAnvil = layoutSystemType === LayoutSystemTypes.ANVIL;
+  const className = isAnvil ? "contents" : "h-full";
 
   const handleMouseLeave: React.MouseEventHandler<HTMLDivElement> = (e) => {
     const wrapperElement = document.getElementById(LAYOUT_WRAPPER_ID);
 
     if (
       isInSideBySideEditor &&
-      (layoutSystemType === LayoutSystemTypes.ANVIL ||
+      (isAnvil ||
         (wrapperElement &&
+          wrapperElement !== e.relatedTarget &&
           e.relatedTarget instanceof Element &&
-          e.relatedTarget.contains(wrapperElement)))
+          e.relatedTarget.contains(wrapperElement) &&
+          e.target instanceof Element))
     ) {
       dispatch(sendAnalyticsForSideBySideHover());
     }
   };
   return (
     <div
-      className="contents"
+      className={className}
       id={LAYOUT_WRAPPER_ID}
       onMouseLeave={handleMouseLeave}
     >
